backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
celery>=5.3.6
redis>=5.0.4
google-genai>=0.8.0
matplotlib>=3.9.0
shapely>=2.0.6
python-dotenv>=1.0.1
pydantic>=2.7.1
Pillow>=10.3.0
python-multipart>=0.0.9
httpx>=0.27.0
numpy>=1.26.4
